+++ CinemaCat.Application/Interfaces/IBlobServiceProvider.cs
@@ -2,7 +2,7 @@
 {
     public interface IBlobServiceProvider
     {
-        Task<Guid> UploadAsync(Stream stream);
+        Task<Guid> UploadAsync(Stream fullImageStream, Stream? previewImageStream = null);
         Task<Stream> DownloadAsync(Guid id, bool preview = false);
     }
 }
