+++ CinemaCat.Infrastructure/Data/AzureBlobProvider.cs
@@ -2,9 +2,6 @@
 using Azure.Storage.Blobs.Models;
 using Azure.Storage.Blobs.Specialized;
 using CinemaCat.Application.Interfaces;
-using SixLabors.ImageSharp;
-using SixLabors.ImageSharp.Formats.Jpeg;
-using SixLabors.ImageSharp.Processing;
 
 namespace CinemaCat.Infrastructure.Data
 {
@@ -23,26 +20,15 @@ namespace CinemaCat.Infrastructure.Data
             return await blob.OpenReadAsync();
         }
 
-        public async Task<Guid> UploadAsync(Stream fileStream)
+        public async Task<Guid> UploadAsync(Stream fullImageStream, Stream? previewImageStream = null)
         {
             var guid = Guid.NewGuid();
             BlobContainerClient containerClient = _blobServiceClient.GetBlobContainerClient("images");
-            using (var fullImageStream = new MemoryStream())
-            using (var previewImageStream = new MemoryStream())
-            using (var image = Image.Load(fileStream))
-            {
-                image.Save(fullImageStream, new JpegEncoder());
-                var ratio = image.Height / image.Width;
-                var clone = image.Clone(x => x.Resize(250 * ratio, 250));
-                clone.Save(previewImageStream, new JpegEncoder());
 
-                previewImageStream.Position = 0;
-                fullImageStream.Position = 0;
-                var blob = containerClient.GetBlockBlobClient("full/" + guid);
-                var blobCompressed = containerClient.GetBlockBlobClient("compressed/" + guid);
-                await blob.UploadAsync(fullImageStream, new BlobHttpHeaders { ContentType = "image/jpeg" });
-                await blobCompressed.UploadAsync(previewImageStream, new BlobHttpHeaders { ContentType = "image/jpeg" });
-            }
+            var blob = containerClient.GetBlockBlobClient("full/" + guid);
+            var blobCompressed = containerClient.GetBlockBlobClient("compressed/" + guid);
+            await blob.UploadAsync(fullImageStream, new BlobHttpHeaders { ContentType = "image/jpeg" });
+            await blobCompressed.UploadAsync(previewImageStream, new BlobHttpHeaders { ContentType = "image/jpeg" });
             return guid;
         }
     }
